Some fixes for engine extensions developed as git projects
- Zip files based on git ls-tree. This also avoids zipping the .git folder, which was previously included.
- Don't allow deleting git project folder when trying to install a local engine extension.

+++ src/com/polydes/repman/LocalRepoBackend.java
@@ -235,6 +235,11 @@ public class LocalRepoBackend extends RepoBackend
 		
 		if(isEngine)
 		{
+			if(new File(installed, ".git").exists())
+			{
+				System.out.println("Can't delete git project -- " + installed.getAbsolutePath());
+				return;
+			}
 			try
 			{
 				FileUtils.deleteDirectory(installed);
